target-build: filter out CONFIG_TARGET_DEVICE config options
This is needed in order to make overriding of the default selection
state for Devices working.

We're downloading config seed which has enabled all the devices already
in the config which makes additional DEFAULT:=n inefficient, so by
deleting all the preconfigured devices from the config seed we let
defconfig do its job and make disabling of the devices actually working.

Signed-off-by: Petr Štetiar <ynezz@true.cz>